move notif count code to navbar component
